refactor(api): move auditLoggerRepository to shared
