+++ api/src/shared/infrastructure/repositories/audit-logger-repository.js
@@ -1,6 +1,6 @@
-import { config } from '../../../shared/config.js';
-import { AuditLoggerApiError } from '../../../shared/domain/errors.js';
-import { httpAgent } from '../../../shared/infrastructure/http-agent.js';
+import { config } from '../../config.js';
+import { AuditLoggerApiError } from '../../domain/errors.js';
+import { httpAgent } from '../http-agent.js';
 
 const { auditLogger } = config;
 const basicAuthorizationToken = btoa(`pix-api:${auditLogger.clientSecret}`);
