+++ api/tests/shared/unit/infrastructure/repositories/audit-logger-repository.test.js
@@ -1,12 +1,12 @@
-import { auditLoggerRepository } from '../../../../../src/identity-access-management/infrastructure/repositories/audit-logger-repository.js';
 import { config } from '../../../../../src/shared/config.js';
 import { AuditLoggerApiError } from '../../../../../src/shared/domain/errors.js';
 import { httpAgent } from '../../../../../src/shared/infrastructure/http-agent.js';
+import { auditLoggerRepository } from '../../../../../src/shared/infrastructure/repositories/audit-logger-repository.js';
 import { catchErr, expect, sinon } from '../../../../test-helper.js';
 
 const { auditLogger } = config;
 
-describe('Unit | Infrastructure | Repositories | audit-logger-repository', function () {
+describe('Unit | Shared | Infrastructure | Repositories | audit-logger-repository', function () {
   let clock;
 
   beforeEach(function () {
